+++ WorldEditCommands/WorldEditCommands.cs
@@ -11,7 +11,7 @@ public class WorldEditCommands : BaseUnityPlugin
 {
   public const string GUID = "world_edit_commands";
   public const string NAME = "World Edit Commands";
-  public const string VERSION = "1.71";
+  public const string VERSION = "1.72";
   public void Awake()
   {
     new Harmony(GUID).PatchAll();